- Updated README with information about how to report bugs - More Work - Fixed sprintf_hl - First try to fix signed/unsigned arithmetic (now i'm using the sign of the left expression, but I don't know how it really works, maybe it checks right too and uses both with some criteria?) - Reduced huge stack trace on CILConverter and shows a tip of where was the last generating sentence - Splitted IntegerExpression into IntegerExpression and CharExpression - More checks and small fixes

+++ ilcc/ilcclib.Tests/Converter/CIL/CILConverterTest.cs
@@ -833,7 +833,9 @@ namespace ilcclib.Tests.Converter.CIL
 				}
 			");
 
-			Assert.AreEqual(unchecked(((uint)0x80000000) >> 16), (uint)(int)Program.GetMethod("test_unsigned").Invoke(null, new object[] { }));
+			//Console.WriteLine(unchecked(((int)0x80000000) >> 16));
+
+			Assert.AreEqual((uint)unchecked(((uint)0x80000000) >> 16), (uint)(int)Program.GetMethod("test_unsigned").Invoke(null, new object[] { }));
 			Assert.AreEqual(unchecked(((int)0x80000000) >> 16), (int)Program.GetMethod("test_signed").Invoke(null, new object[] { }));
 		}
